+++ index.js
@@ -24,7 +24,7 @@ function Mitm() {
 
 Mitm.prototype.on = EventEmitter.prototype.on
 Mitm.prototype.once = EventEmitter.prototype.once
-Mitm.prototype.off = EventEmitter.prototype.off
+Mitm.prototype.off = EventEmitter.prototype.removeListener
 Mitm.prototype.removeListener = EventEmitter.prototype.removeListener
 Mitm.prototype.emit = EventEmitter.prototype.emit